+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `Beaker.workspace.secrets()`.
+- Added `Beaker.secret.get()`.
+- Added `Beaker.secret.read()`.
+- Added `Beaker.secret.write()`.
+- Added `Beaker.secret.delete()`.
+
 ## [v0.6.0](https://github.com/allenai/beaker-py/releases/tag/v0.6.0) - 2022-04-11
 
 ### Added